fix: box-shadow hack for KDE Firefox border corner
This commit changes decoration, decoration backdrop, and chromium, so
that the primary-color border setup, which adds a white border around
the windows, uses a box-shadow rather than a bonafide border.

Despite the assumption that the bottom corners are rectangular, Firefox
on KDE adapts to this even if the about:config setting for GTK bottom
corner rounding is set to true.
However, desktop environments generally did not adapt (tested on Mint).
- GNOME experienced incorrect corners on Firefox and other apps.
- Mint experienced wrongly rectangular bottom corners on other apps.
- XFCE adapted. No issues on xfce4-terminal nor Chromium nor Firefox.